service/requirements.txt
Shapely==1.6.4.post2
shapely-geojson==0.0.1
Flask==1.0.2
Rtree==0.8.3
requests==2.20.1
transit-python==0.8.302
